requirements.txt
# streamlit==1.26.0
# langchain==0.0.354
# chromadb==0.3.29
# sentence-transformers==2.2.2
# pdfplumber==0.9.0
# python-docx==0.8.11
# python-pptx==0.6.21
# tqdm==4.65.0
# openai==1.9.0

streamlit
langchain>=0.0.200
chromadb
sentence-transformers
transformers
# torch           # if using local models; required by sentence-transformers
python-docx
pypdf
pdf2image
pillow
pytesseract
python-magic
cryptography
python-dotenv
openai           # if you use OpenAI
streamlit-authenticator
